safari-600.7-branch/LayoutTests/fast/mediastream/MediaStreamTrack-onended-expected.txt

Tests MediaStreamTrack::onended callback.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



PASS gotStream was called.
PASS track.readyState == "ended" is false
PASS trackClone.readyState == "ended" is false

PASS streamEnded was called.
PASS track.readyState == "ended" is true
PASS trackClone.readyState == "ended" is true

PASS gotStream2 was called.
PASS track.readyState == "ended" is false
PASS trackClone.readyState == "ended" is false

PASS streamEnded2 was called.
PASS track.readyState == "ended" is true
PASS trackClone.readyState == "ended" is true
PASS successfullyParsed is true

TEST COMPLETE
